requirements.txt
 fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
mysql-connector-python
